more changes to support 20-byte serial numbers

+++ proto/scmf.h
@@ -201,7 +201,7 @@ extern int setflagsscm(
 extern int updateblobscm(
     scmcon * conp,
     scmtab * tabp,
-    unsigned long long *snlist,
+    uint8_t *snlist,
     unsigned int sninuse,
     unsigned int snlen,
     unsigned int lid);
